exercise-03/test-fast-api/fast_api_challenge-ml-ae/pyproject.toml
[project]
name = "fast-api-challenge-ml-ae"
version = "0.1.0"
description = ""
authors = [
    {name = "Paloma Ribeiro",email = "paloma.bareli@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)"
]

[tool.poetry]
packages = [{include = "fast_api_challenge_ml_ae", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
